+++ recipes/adfcosmorscompound.py
@@ -1,17 +1,21 @@
 import os, shutil
 from collections import OrderedDict
-from typing import List, Optional, Union, Dict, Mapping
+from typing import List, Optional, Union, Dict, Literal
 
 from scm.plams.interfaces.adfsuite.ams import AMSJob
 from scm.plams.interfaces.adfsuite.crs import CRSJob
-from scm.plams.tools.kftools import KFFile
+from scm.plams.interfaces.adfsuite.densf import DensfJob
+
 from scm.plams.tools.periodic_table import PeriodicTable
 from scm.plams.mol.molecule import Molecule
 from scm.plams.core.basejob import MultiJob
 from scm.plams.core.results import Results
 from scm.plams.core.settings import Settings
-from scm.plams.core.functions import add_to_instance
+from scm.plams.core.functions import add_to_instance, requires_optional_package
 from scm.plams.interfaces.adfsuite.quickjobs import model_to_settings
+
+from scm.plams.tools.hbc_utilities import parse_mesp, write_HBC_to_COSKF, view_HBC
+import numpy as np
 
 __all__ = ["ADFCOSMORSCompoundJob", "ADFCOSMORSCompoundResults"]
 
@@ -58,9 +62,10 @@ class ADFCOSMORSCompoundJob(MultiJob):
         coskf_dir  : The directory in which to place the generated .coskf file.  If nothing is specified, the file will be put in the plams directory corresponding to the job.
         preoptimization  : If None, do not preoptimize with a fast engine (then initial optimization is done with ADF). Otherwise, can be one of 'UFF', 'GAFF', 'GFNFF', 'GFN1-xTB', 'ANI-2x'. Note that you need valid licenses for ForceField or DFTB or MLPotential to use these preoptimizers.
         singlepoint (bool) :  Run a singlepoint in gasphase and with solvation to generate the .coskf file on the given Molecule. (no geometry optimization). Cannot be combined with ``preoptimization``.
+        densf2hbc (bool) : Defaults to True. Performs DENSF analysis to determine the hydrogen bond center (HBC) used in COSMOSAC-DHB-MESP.
         settings (Settings) : A |Settings| object.  settings.runscript.nproc, settings.input.adf.custom_options. If 'adf' is in settings.input it should be provided without the solvation block.
-        name : an optional name for the calculation directory
         mol_info (dict) : an optional dictionary containing information will be written to the Compound Data section within the COSKF file.
+        name : an optional name for the calculation directory
 
     Example:
 
@@ -90,7 +95,8 @@ class ADFCOSMORSCompoundJob(MultiJob):
         preoptimization: Optional[str] = None,
         singlepoint: bool = False,
         settings: Optional[Settings] = None,
-        mol_info: Optional[Mapping[str, Union[float, int, str]]] = None,
+        mol_info: Optional[Dict[str, Union[float, int, str]]] = None,
+        densf2hbc: bool = True,
         **kwargs,
     ):
         """
@@ -120,7 +126,6 @@ class ADFCOSMORSCompoundJob(MultiJob):
         self.mol_info = dict()
         if mol_info is not None:
             self.mol_info.update(mol_info)
-        # self.mol_info = mol_info
         self.atomic_ion = False  # should be set when molecule is set if using a custom prerun() method
         if molecule is not None:
             self.mol_info["Molar Mass"] = molecule.get_mass()
@@ -138,20 +143,23 @@ class ADFCOSMORSCompoundJob(MultiJob):
 
         self.coskf_name = coskf_name
         self.coskf_dir = coskf_dir
+        self.densf2hbc = densf2hbc
 
         if self.coskf_dir is not None and not os.path.exists(self.coskf_dir):
             os.mkdir(self.coskf_dir)
 
-        if self.coskf_name is not None and isinstance(self.coskf_name, str) and not self.coskf_name.endswith(".coskf"):
-            self.coskf_name += ".coskf"
-        elif self.coskf_name is None:
+        if self.coskf_name is None:
             self.coskf_name = f"{self.name}.coskf"
+        elif isinstance(self.coskf_name, str) and not self.coskf_name.endswith(".coskf"):
+            self.coskf_name += ".coskf"
 
         gas_s = Settings()
-        gas_s += self.adf_settings(solvation=False, settings=self.settings)
+        gas_s += ADFCOSMORSCompoundJob.adf_settings(solvation=False, settings=self.settings)
         gas_job = AMSJob(settings=gas_s, name="gas")
 
         if not singlepoint:
+            gas_job.settings.input.ams.Task = "GeometryOptimization"
+
             if preoptimization:
                 preoptimization_s = Settings()
                 preoptimization_s.runscript.nproc = 1
@@ -162,13 +170,6 @@ class ADFCOSMORSCompoundJob(MultiJob):
                 )
                 self.children["preoptimization"] = preoptimization_job
 
-            gas_s = Settings()
-            gas_s.input.ams.Task = "GeometryOptimization"
-            gas_s += self.adf_settings(solvation=False, settings=self.settings)
-            gas_job = AMSJob(settings=gas_s, name="gas")
-
-            if preoptimization:
-
                 @add_to_instance(gas_job)
                 def prerun(self):  # noqa: F811
                     self.molecule = self.parent.children["preoptimization"].results.get_main_molecule()
@@ -176,21 +177,24 @@ class ADFCOSMORSCompoundJob(MultiJob):
             else:
                 gas_job.molecule = self.input_molecule
 
-            self.children["gas"] = gas_job
+        elif singlepoint:
+            gas_job.settings.input.ams.Task = "SinglePoint"
+            gas_job.molecule = self.input_molecule
+
+        self.children["gas"] = gas_job
+
+        if self.densf2hbc:
+            densf_job = DensfJob(settings=ADFCOSMORSCompoundJob.densf_settings(), name="densf")
+            self.children["densf"] = densf_job
+
+            @add_to_instance(densf_job)
+            def prerun(self):  # noqa: F811
+                gas_job.results.wait()
+                self.inputjob = f"../gas/adf.rkf #{self.parent.name}"
 
         solv_s = Settings()
         solv_s.input.ams.Task = "SinglePoint"
         solv_job = AMSJob(settings=solv_s, name="solv")
-
-        if singlepoint:
-
-            gas_job.settings.input.ams.Task = "SinglePoint"
-
-            @add_to_instance(gas_job)
-            def prerun(self):  # noqa: F811
-                self.molecule = self.parent.input_molecule
-
-            self.children["gas"] = gas_job
 
         @add_to_instance(solv_job)
         def prerun(self):  # noqa: F811
@@ -200,26 +204,27 @@ class ADFCOSMORSCompoundJob(MultiJob):
             molecule_charge = gas_job.results.get_main_molecule().properties.get("charge", 0)
             self.settings.input.ams.LoadSystem._1 = f"# {self.parent.name}"
             self.settings.input.ams.LoadSystem._2 = f"# charge {molecule_charge}"
-            self.settings += self.parent.adf_settings(
+            self.settings += ADFCOSMORSCompoundJob.adf_settings(
                 solvation=True,
                 settings=self.parent.settings,
                 elements=list(set(at.symbol for at in self.parent.input_molecule)),
                 atomic_ion=self.parent.atomic_ion,
             )
-            # self.settings.input.ams.EngineRestart = self.parent.children['gas'].results.rkfpath(file='adf') # this doesn't work with PLAMS restart since the file will refer to the .res directory (so the job is rerun needlessly)
-            # self.settings.input.ams.LoadSystem.File = self.parent.children['gas'].results.rkfpath(file='ams')
-            # cannot copy to gasphase-ams.rkf etc. because that conflicts with PLAMS restarts
-            # shutil.copyfile(gas_job.results.rkfpath(file='ams'), os.path.join(self.path, 'gasphase-ams.rkf'))
-            # shutil.copyfile(gas_job.results.rkfpath(file='adf'), os.path.join(self.path, 'gasphase-adf.rkf'))
 
         @add_to_instance(solv_job)
         def postrun(self):
-            self.parent.convert_to_coskf(
-                self.results.rkfpath(file="adf"),
-                self.parent.coskf_name,
-                self.parent.path,
-                self.parent.coskf_dir,
-                self.parent.mol_info,
+            if self.parent.densf2hbc:
+                densf_job.results.wait()
+                densf_path = densf_job.results.kfpath()
+            else:
+                densf_path = None
+            ADFCOSMORSCompoundJob.convert_to_coskf(
+                rkf_path=self.results.rkfpath(file="adf"),
+                coskf_name=self.parent.coskf_name,
+                plams_dir=self.parent.path,
+                coskf_dir=self.parent.coskf_dir,
+                mol_info=self.parent.mol_info,
+                densf_path=densf_path,
             )
 
         self.children["solv"] = solv_job
@@ -428,39 +433,81 @@ class ADFCOSMORSCompoundJob(MultiJob):
         return s
 
     @staticmethod
+    def densf_settings(grid: Literal["Medium", "Fine"] = "Medium") -> Settings:
+        s = Settings()
+        s.input.GRID = f"{grid}\nEnd"
+        s.input.Density = "SCF"
+        s.input.Potential = "COUL SCF"
+        return s
+
+    @staticmethod
+    @requires_optional_package("scm.libbase")
     def convert_to_coskf(
-        rkf: str,
+        rkf_path: str,
         coskf_name: str,
         plams_dir: str,
         coskf_dir: Optional[str] = None,
-        mol_info: Optional[Mapping[str, Union[float, int, str]]] = None,
-    ):
+        mol_info: Optional[Dict[str, Union[float, int, str]]] = None,
+        densf_path: Optional[str] = None,
+    ) -> None:
         """
-        rkf: str
-            absolute path to adf.rkf
+        Convert an adf.rkf file into a .coskf file
 
-        coskf_name: str
-            the name of the .coskf file
-
-        plams_dir: str
-            plamsjob path to write out the .coskf file
-
-        coskf_dir: Optional[str]
-            additional path to store the .coskf file
-
-        mol_info: dict[str, float | int | str], optional
-            Optional information to write out in the "Compound Data" section of the .coskf file
-
+        Args:
+            rkf_path (str) : absolute path to adf.rkf
+            coskf_name (str) : the name of the .coskf file
+            plams_dir (str) : plamsjob path to write out the .coskf file
+            coskf_dir (Optional[str]) :additional path to store the .coskf file
+            mol_info (Optional[Dict[str, Union[float, int, str]]]) : Optional information to write out in the "Compound Data" section of the .coskf file
+            densf_path (Optional[str]) : path to the densf output .t41 file
         """
-        f = KFFile(rkf)
-        cosmo = f.read_section("COSMO")
-        coskf_file = KFFile(os.path.join(plams_dir, coskf_name), autosave=False)
-        for k, v in cosmo.items():
-            coskf_file.write("COSMO", k, v)
-        if mol_info is not None:
+        from scm.libbase import KFFile
+
+        with KFFile(rkf_path) as rkf:
+            cosmo = rkf.read_section("COSMO")
+
+        coskf_path = os.path.join(plams_dir, coskf_name)
+
+        with KFFile(coskf_path, autosave=False) as rkf:
+            for key, value in cosmo.items():
+                rkf.write("COSMO", key, float(value) if isinstance(value, np.float64) else value)
             for key, value in mol_info.items():
-                coskf_file.write("Compound Data", key, value)
-        coskf_file.save()
+                rkf.write("Compound Data", key, float(value) if isinstance(value, np.float64) else value)
+
+        if densf_path is not None:
+            HBC_xyz, HBC_atom, HBC_angle, HBC_info = parse_mesp(densf_path, coskf_path)
+            write_HBC_to_COSKF(coskf_path, HBC_xyz, HBC_atom, HBC_angle, HBC_info)
 
         if coskf_dir is not None:
-            shutil.copy2(os.path.join(plams_dir, coskf_name), os.path.join(coskf_dir, coskf_name))
+            shutil.copy2(coskf_path, os.path.join(coskf_dir, coskf_name))
+
+    @staticmethod
+    def update_hbc_to_coskf(coskf: str, visulization: bool = False) -> None:
+        """
+        Determine the hydrogen bond center for existing COSKF file
+
+        Args:
+            coskf (str) : Existing COSKF file
+            visulization (bool) : Visulization of hydrogen bond center
+        """
+        molecule = Molecule(coskf)
+        coskf_name = os.path.basename(coskf).replace(".coskf", "")
+
+        atomic_ion = len(molecule.atoms) == 1
+        gas_settings = ADFCOSMORSCompoundJob.adf_settings(solvation=False, atomic_ion=atomic_ion)
+        gas_settings.input.ams.Task = "SinglePoint"
+        gas_job = AMSJob(molecule=molecule, settings=gas_settings, name=f"gas_{coskf_name}")
+        gas_job.run()
+
+        gas_rkf = gas_job.results.rkfpath(file="adf")
+        densf_settings = ADFCOSMORSCompoundJob.densf_settings()
+        densf_job = DensfJob(gas_rkf, settings=densf_settings, name=f"densf_{coskf_name}")
+        densf_job.run()
+
+        t41 = densf_job.results.kfpath()
+
+        HBC_xyz, HBC_atom, HBC_angle, HBC_info = parse_mesp(t41, coskf)
+        write_HBC_to_COSKF(coskf, HBC_xyz, HBC_atom, HBC_angle, HBC_info)
+
+        if visulization:
+            view_HBC(coskf)