+++ CHANGELOG.md
@@ -31,6 +31,8 @@ This changelog is effective from the 2025 releases.
 * Methods `Settings.nested_keys`, `Settings.block_keys` for accessing nested keys and `Settings.contains_nested` and `Settings.pop_nested` for checking if nested keys exist in a settings object and popping them
 * Method `Settings.compare` added to compare and contrast items in two settings objects
 * Method `readcoskf` added to `Molecule` class, enabling the reading of COSKF file
+* Added support for calculating the hydrogen bond center using the Densf calculation in the `ADFCOSMORSCompoundJob` class
+* Introduced the `update_hbc_to_coskf` method in `ADFCOSMORSCompoundJob` class to calculate the hydrogen bond center using an existing COSKF file
 
 ### Changed
 * Functions for optional packages (e.g. RDKit, ASE) are available even when these packages are not installed, but will raise an `MissingOptionalPackageError` when called